# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AlgoRace API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  name: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  id: number;
  email: string;
  name: string;
  elo: number;
  wins: number;
  losses: number;
  createdAt: string;
}

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  elo: number;
  wins: number;
  losses: number;
  createdAt: string;
}

export type ProblemSummaryDifficulty =
  (typeof ProblemSummaryDifficulty)[keyof typeof ProblemSummaryDifficulty];

export const ProblemSummaryDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface ProblemSummary {
  id: number;
  title: string;
  slug: string;
  difficulty: ProblemSummaryDifficulty;
  tags: string[];
}

export interface ProblemExample {
  input: string;
  output: string;
  /** @nullable */
  explanation: string | null;
}

export type ProblemDifficulty =
  (typeof ProblemDifficulty)[keyof typeof ProblemDifficulty];

export const ProblemDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Problem {
  id: number;
  title: string;
  slug: string;
  description: string;
  difficulty: ProblemDifficulty;
  tags: string[];
  constraints: string;
  examples: ProblemExample[];
  starterCodeJs: string;
  starterCodePy: string;
}

export type RunCodeBodyLanguage =
  (typeof RunCodeBodyLanguage)[keyof typeof RunCodeBodyLanguage];

export const RunCodeBodyLanguage = {
  javascript: "javascript",
  python: "python",
} as const;

export interface RunCodeBody {
  language: RunCodeBodyLanguage;
  code: string;
}

export interface TestCaseResult {
  passed: boolean;
  input: string;
  expectedOutput: string;
  actualOutput: string;
  /** @nullable */
  error: string | null;
}

export interface RunResult {
  passed: number;
  total: number;
  results: TestCaseResult[];
}

export interface CreateMatchBody {
  /** @nullable */
  problemId?: number | null;
}

export interface MatchPlayer {
  userId: number;
  name: string;
  elo: number;
  /** @nullable */
  testsPassedCount: number | null;
  /** @nullable */
  totalTests: number | null;
  /** @nullable */
  language: string | null;
  /** @nullable */
  submittedAt: string | null;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  waiting: "waiting",
  active: "active",
  finished: "finished",
  abandoned: "abandoned",
} as const;

export interface Match {
  id: number;
  status: MatchStatus;
  /** @nullable */
  inviteCode: string | null;
  problemId: number;
  problem: Problem;
  players: MatchPlayer[];
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  endedAt: string | null;
  createdAt: string;
}

export interface MatchSummary {
  id: number;
  status: string;
  problemTitle: string;
  problemDifficulty: string;
  /** @nullable */
  opponentName: string | null;
  /** @nullable */
  opponentElo: number | null;
  /** @nullable */
  won: boolean | null;
  /** @nullable */
  eloChange: number | null;
  /** @nullable */
  duration: number | null;
  createdAt: string;
}

export type SubmitCodeBodyLanguage =
  (typeof SubmitCodeBodyLanguage)[keyof typeof SubmitCodeBodyLanguage];

export const SubmitCodeBodyLanguage = {
  javascript: "javascript",
  python: "python",
} as const;

export interface SubmitCodeBody {
  matchId: number;
  language: SubmitCodeBodyLanguage;
  code: string;
}

export interface SubmissionResult {
  passed: number;
  total: number;
  allPassed: boolean;
  won: boolean;
  /** @nullable */
  eloChange: number | null;
  results: TestCaseResult[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  elo: number;
  wins: number;
  losses: number;
  winRate: number;
  totalMatches: number;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  elo: number;
  wins: number;
  losses: number;
  totalMatches: number;
  winRate: number;
  createdAt: string;
}

export interface EloHistoryEntry {
  matchId: number;
  oldElo: number;
  newElo: number;
  createdAt: string;
}

export interface QueueStatus {
  inQueue: boolean;
  queueSize: number;
  /** @nullable */
  matchId: number | null;
}

export interface Friend {
  userId: number;
  name: string;
  email: string;
  elo: number;
  wins: number;
  losses: number;
  friendshipId: number;
  since: string;
}

export interface FriendRequest {
  id: number;
  requesterId: number;
  requesterName: string;
  requesterEmail: string;
  requesterElo: number;
  createdAt: string;
}

export type UserSearchResultRelationship =
  (typeof UserSearchResultRelationship)[keyof typeof UserSearchResultRelationship];

export const UserSearchResultRelationship = {
  none: "none",
  pending_outgoing: "pending_outgoing",
  pending_incoming: "pending_incoming",
  friends: "friends",
  self: "self",
} as const;

export interface UserSearchResult {
  id: number;
  name: string;
  email: string;
  elo: number;
  relationship: UserSearchResultRelationship;
}

export interface SendFriendRequestBody {
  userId: number;
}

export interface ChallengeFriendBody {
  /** @nullable */
  problemId?: number | null;
}

export interface PlatformStats {
  activePlayers: number;
  totalMatches: number;
  totalProblems: number;
  matchesToday: number;
}

export type ListProblemsParams = {
  difficulty?: ListProblemsDifficulty;
  tag?: string;
  search?: string;
};

export type ListProblemsDifficulty =
  (typeof ListProblemsDifficulty)[keyof typeof ListProblemsDifficulty];

export const ListProblemsDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type SearchUsersParams = {
  q: string;
};