# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AlgoRace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /friends/requests:
    get:
      operationId: listFriendRequests
      tags: [friends]
      summary: List the current user's pending incoming friend requests
      responses:
        "200":
          description: Friend request list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"

    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFriendRequestBody"
      responses:
        "201":
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Friendship already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/requests/{id}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a pending friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friend request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/requests/{id}/decline:
    post:
      operationId: declineFriendRequest
      tags: [friends]
      summary: Decline a pending friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Friend request declined
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/search:
    get:
      operationId: searchUsers
      tags: [friends]
      summary: Search for users by name or email to add as friends
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchResult"
  /friends/{userId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Friend removed
        "404":
          description: Friendship not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/{userId}/challenge:
    post:
      operationId: challengeFriend
      tags: [friends]
      summary: Challenge a friend to a direct match
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeFriendBody"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not friends with this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
      required:
        - email
        - name
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        elo:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        createdAt:
          type: string
      required:
        - token
        - id
        - email
        - name
        - elo
        - wins
        - losses
        - createdAt
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        elo:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - elo
        - wins
        - losses
        - createdAt
...
    ProblemExample:
      type: object
      properties:
        input:
          type: string
        output:
          type: string
        explanation:
          type: ["string", "null"]
      required:
        - input
        - output
        - explanation
    Problem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
        constraints:
          type: string
        examples:
          type: array
          items:
            $ref: "#/components/schemas/ProblemExample"
        starterCodeJs:
          type: string
        starterCodePy:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - difficulty
        - tags
        - constraints
        - examples
        - starterCodeJs
        - starterCodePy
...
    MatchPlayer:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        elo:
          type: integer
        testsPassedCount:
          type: ["integer", "null"]
        totalTests:
          type: ["integer", "null"]
        language:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
      required:
        - userId
        - name
        - elo
        - testsPassedCount
        - totalTests
        - language
        - submittedAt
    Match:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [waiting, active, finished, abandoned]
        inviteCode:
          type: ["string", "null"]
        problemId:
          type: integer
        problem:
          $ref: "#/components/schemas/Problem"
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        winnerId:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - status
        - inviteCode
        - problemId
        - problem
        - players
        - winnerId
        - startedAt
        - endedAt
        - createdAt
...
    Friend:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        elo:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        friendshipId:
          type: integer
        since:
          type: string
      required:
        - userId
        - name
        - email
        - elo
        - wins
        - losses
        - friendshipId
        - since
    FriendRequest:
      type: object
      properties:
        id:
          type: integer
        requesterId:
          type: integer
        requesterName:
          type: string
        requesterEmail:
          type: string
        requesterElo:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - requesterId
        - requesterName
        - requesterEmail
        - requesterElo
        - createdAt
    UserSearchResult:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        elo:
          type: integer
        relationship:
          type: string
          enum: [none, pending_outgoing, pending_incoming, friends, self]
      required:
        - id
        - name
        - email
        - elo
        - relationship
    SendFriendRequestBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
    ChallengeFriendBody:
      type: object
      properties:
        problemId:
          type: ["integer", "null"]
      required: []
...